refactor: api client (#51)
