+++ src/services/VisionApi/Workspace.js
@@ -1,61 +1,32 @@
-const axios = require("axios");
+import ApiClient from "./ApiClient";
 
-import formatErrors from "./errors";
+export default class WorkspaceService extends ApiClient {
+  static findAll(userToken, params) {
+    let url = params ? `workspaces${params}` : `workspaces`;
 
-export default class WorkspaceService {
-  static async findAll(userToken, params) {
-    let response = null;
-    let errors = null;
-
-    try {
-      response = await axios.get(
-        params ? `workspaces${params}` : `workspaces`,
-        {
-          headers: { Authorization: `Bearer ${userToken}` },
-        }
-      );
-    } catch (apiErrors) {
-      errors = formatErrors(apiErrors.response.data.errors);
-    }
-
-    return { response, errors };
+    return this.makeRequest("GET", url, null, {
+      Authorization: `Bearer ${userToken}`,
+    });
   }
 
   static async create(userToken, data) {
-    let response = null;
-    let errors = null;
-
-    try {
-      response = await axios.post(
-        `workspaces`,
-        {
-          name: data.name,
-          logo: data.logo,
-          color: data.color,
-        },
-        {
-          headers: { Authorization: `Bearer ${userToken}` },
-        }
-      );
-    } catch (apiErrors) {
-      errors = formatErrors(apiErrors.response.data.errors);
-    }
-
-    return { response, errors };
+    return this.makeRequest(
+      "POST",
+      `workspaces`,
+      {
+        name: data.name,
+        logo: data.logo,
+        color: data.color,
+      },
+      {
+        Authorization: `Bearer ${userToken}`,
+      }
+    );
   }
 
   static async find(userToken, id) {
-    let response = null;
-    let errors = null;
-
-    try {
-      response = await axios.get(`workspaces/${id}`, {
-        headers: { Authorization: `Bearer ${userToken}` },
-      });
-    } catch (apiErrors) {
-      errors = formatErrors(apiErrors.response.data.errors);
-    }
-
-    return { response, errors };
+    return this.makeRequest("GET", `workspaces/${id}`, null, {
+      Authorization: `Bearer ${userToken}`,
+    });
   }
 }
